requirements.txt
numpy==1.18.1
pycaret==2.1.2
pandas==1.0.1
streamlit==0.64.0
Pillow==7.2.0
